refactor logic into separate files
